+++ Classes/Frontend/View/TeamListView.php
@@ -1,12 +1,22 @@
 <?php
 
+namespace System25\T3sports\Frontend\View;
+
+use Exception;
+use Sys25\RnBase\Frontend\Marker\BaseMarker;
+use Sys25\RnBase\Frontend\Marker\Templates;
+use Sys25\RnBase\Frontend\Request\RequestInterface;
+use Sys25\RnBase\Frontend\View\ContextInterface;
+use Sys25\RnBase\Frontend\View\Marker\BaseView;
 use Sys25\RnBase\Maps\Factory;
 use System25\T3sports\Utility\MapsUtil;
+use tx_cfcleaguefe_util_TeamMarker as TeamMarker;
+use tx_rnbase;
 
 /***************************************************************
  *  Copyright notice
  *
- *  (c) 2007-2017 Rene Nitzsche (rene@system25.de)
+ *  (c) 2007-2021 Rene Nitzsche (rene@system25.de)
  *  All rights reserved
  *
  *  This script is part of the TYPO3 project. The TYPO3 project is
@@ -25,38 +35,35 @@ use System25\T3sports\Utility\MapsUtil;
  *
  *  This copyright notice MUST APPEAR in all copies of the script!
  ***************************************************************/
-tx_rnbase::load('tx_rnbase_view_Base');
-tx_rnbase::load('tx_rnbase_util_ListBuilder');
-tx_rnbase::load('tx_rnbase_maps_Factory');
-tx_rnbase::load('tx_rnbase_util_Templates');
 
 /**
  * Viewklasse für die Anzeige einer Teamliste.
  * Isn't it tiny! ;-).
  */
-class tx_cfcleaguefe_views_TeamList extends tx_rnbase_view_Base
+class TeamListView extends BaseView
 {
     /**
      * Erstellen des Frontend-Outputs.
      */
-    public function createOutput($template, &$viewData, &$configurations, &$formatter)
+    public function createOutput($template, RequestInterface $request, $formatter)
     {
+        $viewData = $request->getViewContext();
         // Die ViewData bereitstellen
-        $teams = &$viewData->offsetGet('teams');
+        $teams = $viewData->offsetGet('teams');
         $listBuilder = tx_rnbase::makeInstance('tx_rnbase_util_ListBuilder');
 
-        $template = $listBuilder->render($teams, $viewData, $template, 'tx_cfcleaguefe_util_TeamMarker', 'teamlist.team.', 'TEAM', $formatter);
-        if (tx_rnbase_util_BaseMarker::containsMarker($template, 'TEAMMAP')) {
-            $markerArray['###TEAMMAP###'] = $this->getMap($teams, $configurations, $this->getController()
+        $template = $listBuilder->render($teams, $viewData, $template, TeamMarker::class, 'teamlist.team.', 'TEAM', $formatter);
+        if (BaseMarker::containsMarker($template, 'TEAMMAP')) {
+            $markerArray['###TEAMMAP###'] = $this->getMap($teams, $request->getConfigurations(), $request
                 ->getConfId().'map.', 'TEAM');
         }
 
-        $out = tx_rnbase_util_Templates::substituteMarkerArrayCached($template, $markerArray); // , $wrappedSubpartArray);
+        $out = Templates::substituteMarkerArrayCached($template, $markerArray); // , $wrappedSubpartArray);
 
         return $out;
     }
 
-    public function getMainSubpart(&$viewData)
+    public function getMainSubpart(ContextInterface $viewData)
     {
         return '###TEAM_LIST###';
     }