+++ CHANGELOG.md
@@ -3,7 +3,9 @@ Changes
 -------
 
 v2.0.0 (??.??.2021)
-* action tx_cfcleaguefe_actions_LeagueTableShow removed
+* all actions moved to PSR-4 namespace
+* action `tx_cfcleaguefe_actions_LeagueTableShow` removed
+* all html templates moved to `Resources/Private/Templates/Html/`
 
 v1.7.2 (19.05.2021)
  * Fix matchtable shows home matches only